// src/app/shared/enum/work-order.enum.ts
export enum EWorkOrderStatus {
  ALL = 'ALL',
  OPEN = 'OPEN',
  INSPECTING = 'INSPECTING',
  WAITING_QUOTATION = 'WAITING_QUOTATION',
  WAITING_APPROVAL = 'WAITING_APPROVAL',
  WAITING_ASSIGNMENT = 'WAITING_ASSIGNMENT',
  IN_PROGRESS = 'IN_PROGRESS',
  WAITING_ADDITIONAL_APPROVAL = 'WAITING_ADDITIONAL_APPROVAL',
  WAITING_QC = 'WAITING_QC',
  REWORK = 'REWORK',
  QC_APPROVED = 'QC_APPROVED',
  HOLD = 'HOLD',
  READY_DELIVERY = 'READY_DELIVERY',
  COMPLETED = 'COMPLETED',
  CANCELLED = 'CANCELLED',
}